Swagger added, Refactor UserService to use UserForm instead of User
There were major updates in handling user data. To enhance the encapsulation of the data, a UserForm is now employed to transfer user data for creating a new user. The UserService and UserServiceImpl were updated to facilitate this change. UserForm provides a simpler and more controlled manner of creating user entities. This also led to the creation of UserController to handle user creation requests using UserForms as request bodies. Additionally, UserDTO was introduced to provide a more streamlined representation of user data for easier data transfer and operations. Also, pom.xml is updated to use different dependencies.