Add validation and shipping methods to ParticipationService (Blob foutu)
Enhanced the ParticipationService by adding methods to find validated, non-validated, shipped, and non-shipped participations and methods for validating and shipping participations. Also, the ParticipationForm, ParticipationRepository, ParticipationController and ParticipationServiceImpl were significantly refactored to enhance data handling and improve readability and maintainability. Mimetypes in Participation model are changed to byte arrays to handle image data in a more efficient way. ParticipationForm was shifted to record for compactness and ParticipationDTO was added for easy data transfer.