Add email service and update config files (Pas encore testé)
This commit introduces an EmailService and modifies the ParticipationService to use this new Email Service. In addition, several dependencies related to security and mailing have been added to the pom.xml file. Lastly, a new MailConfiguration class and environment variable changes to the application.yml file have been made to support the email functionality.