Add function to count participation by province
Added a new function to count participation records per province. The function groups records by the postcode of the participants. Also, it modifies the postCode data type from String to int to facilitate this grouping. The added function increases the ability to analyze the distribution of participants, especially for administrative users.