Add method to find all other satisfaction comments
A new method has been introduced to obtain all other satisfaction comments. This is achieved by adding this method in the Participation service, implementing it in the service implementation class, and including the query in the repository. A new endpoint for this functionality has also been created in the Participation controller, accessible by users with 'ADMIN' or 'LOGISTIC' roles.