Add 'countAllOtherProductType' method in Participation service
A new method 'countAllOtherProductType' is added in Participation service for counting items not categorized as 'insecticide', 'herbicide' or 'fongicide'. This method implementation is added in the service interface, service implementation, and the repository interface. The corresponding endpoint in the participation controller is also created, allowing administrators and logistic users to engage with this function.